--- week2/src/App.css
@@ -1,1 +1,0 @@
-
