+++ week2/src/main.jsx
@@ -1,7 +1,6 @@
 import React from 'react'
 import ReactDOM from 'react-dom/client'
-import App from './App.jsx'
-import './index.css'
+import App from './practice_1st'
 
 ReactDOM.createRoot(document.getElementById('root')).render(
   <React.StrictMode>
